AG-43923: Additional tests

+++ crates/ffi/src/outer_error.rs
@@ -71,7 +71,7 @@ pub enum AGOuterError {
 
     /// Filter integrity check failed — signature is missing or does not match
     #[error("Filter integrity check failed for filter_id={0}")]
-    FilterIntegrityCheckFailed(i64),
+    FilterIntegrityCheckFailed(i32),
 
     /// Other errors
     #[error("{0}")]
